clean up comments etc
