+++ py_flask/routes/instructor_routes.py
@@ -140,31 +140,8 @@ def get_logs():
     except Err_Custom_FullInfo as err:
         return err.get_response()
     except Exception as err:
-        # handle unexpected exceptions
         generic_error = Err_Custom_FullInfo({'message': str(err), 'status_code': 500})
         return generic_error.get_response()
-
-
-# @blueprint_instructor.route("/get_logs", methods=['GET'])
-# @jwt_and_csrf_required
-# def get_logs():
-#     instructor_only()
-
-#     try:
-#         logData = getLogs()
-#         if logData is None:
-
-#             newErr = {
-#                 'message': 'error retrieving logs',
-#                 'status_code': 400
-#             }
-#             raise Err_Custom_FullInfo(newErr)
-#         else:
-#             return logData
-#     except:
-#         return Exception
-
-
 
 @blueprint_instructor.route("/get_instructor_data", methods=['GET'])
 @jwt_and_csrf_required
@@ -185,7 +162,6 @@ def get_instructor_data():
     except Err_Custom_FullInfo as err:
         return err.get_response()
     except Exception as err:
-        # handle unexpected exceptions
         generic_error = Err_Custom_FullInfo({'message': str(err), 'status_code': 500})
         return generic_error.get_response()
 
@@ -476,7 +452,7 @@ def clear_groups():
 def get_chat_library():
     instructor_only()
 
-    # chatHistoryData_dict containins: "unordered_messages_list", "user_channels_dict"
+    # chatHistoryData_dict contains: "unordered_messages_list", "user_channels_dict"
     chatHistoryData_dict = getChatLibrary()
     return jsonify(chatHistoryData_dict)
