+++ react/components/Dropdown.jsx
@@ -7,25 +7,11 @@ function Dropdown({
     optionsArray,
     choice_state,
     choice_stateSetter, // function to update state in the parent component
-    validation_setter  // function to update validation state in the parent component
 }) {
-
-    const [localChoice, set_localChoice] = useState('ALL')
-
-    // const validateSelection = (value) => {
-    //     // selection is 'valid' if a choice has been made
-    //     return value != null && value !== "";
-    // };
-
-    // useEffect(() => {
-    //     // validation_setter(validateSelection(choice_state));
-    //     choice_stateSetter(choice_state);
-    // }, [localChoice]);
-
+    
     const handleDropdownChange = (event) => {
         const newValue = event.target.value;
         choice_stateSetter(newValue);
-        // validation_setter(validateSelection(newValue));
     };
 
     return (
