+++ react/pages/student/scenarios/panes/guide/Q_and_A/SubmitButton.jsx
@@ -22,8 +22,6 @@ function SubmitButton({ scenario_id, question_num, scenario_type }) {
                 const scoresArray = evaluated.data.points_gained;
                 const testingItem = scoresArray[0];
                 const pointsAward = testingItem.points_awarded;
-                // dev_fix (don't show answer once working reliably)
-                // console.log(`Submission for scenario ${scenario_id} question ${question_num}: "${inputText_state}". Correct answer was ${testingItem.correct_response}. You were awarded ${pointsAward} points!`)
             }
         } catch (err) {
             console.log(`Submission error: ${err}`);
